+++ types/ObjectType.d.ts
@@ -1,0 +1,13 @@
+import { SchemaDeclaration } from './SchemaDeclaration';
+import { Type } from './Type';
+
+export declare class ObjectType<ValueType = any, DataType = any, ErrorMsgType = string> extends Type<ValueType, DataType, ErrorMsgType> {
+    constructor(errorMessage?: ErrorMsgType);
+    shape: (types: SchemaDeclaration<ValueType, ErrorMsgType>) => this;
+}
+
+declare function getObjectType<ValueType = any, DataType = any, ErrorMsgType = string>(errorMessage?: ErrorMsgType): ObjectType<ValueType, DataType, ErrorMsgType>;
+
+type exportType = typeof getObjectType;
+
+export default exportType;