+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- Adding the typescript types declaration in to package
+
 # 1.3.1
 
 - Fixed an issue where `isOneOf` was not valid in `StringType` (#18)